% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainfuncs.r
\name{clinic_list}
\alias{clinic_list}
\title{clinic_list()}
\usage{
clinic_list(start, end, excluded = NULL, bank_hols = NULL)
}
\arguments{
\item{start}{First day of time period to create events in}

\item{end}{Last day of time period}

\item{excluded}{A date vector containing days to be excluded ie no events to be created on these days}

\item{bank_hols}{A date vector containing the dates of any bank holidays in the time period. No events
will be created on bank holidays}
}
\value{
A data frame containing details of clinics to create
}
\description{
Take a start and end day and find work days
}
\examples{
list_of_clinics <- clinic_list(dmy("1/1/17"), dmy("1/2/17"))
more_clinics <- clinic_list(today(), today()+day(14), bank_hols=scrape_bh_dates())
}
