% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gapi.r
\name{gAPI_authenticate}
\alias{gAPI_authenticate}
\title{gAPI_authenticate()}
\usage{
gAPI_authenticate()
}
\description{
Handles signing in to Google. If necessary will open a browser window to sign in.
}
\details{
Requires a file called \code{credentials.dat} containing a client ID and a client secret. These need
to be obtained from Google - Google API console. Put the client ID in a variable called \code{gAPIclientID}
and the secret in \code{gAPIsecret}. Create the \code{credentials.dat} file with \code{save(gAPIclientID, gAPIsecret, file="credentials.dat")}
}
\examples{
gAPI_authenticate()

}
