% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainfuncs.r
\name{get_existing_appointments}
\alias{get_existing_appointments}
\title{get_existing_appointments()
Get a list of appointments already in calendar to avoid when creating clinics}
\usage{
get_existing_appointments(start = today(), end = today() + days(7))
}
\arguments{
\item{start}{Start of period to search for pre existing appointments in}

\item{end}{End of time period}
}
\value{
A list containing two items, both vectors of dates. \code{$exclude} contains days with
  events that would otherwise clash with clinics. \code{$sundays} contains Sunday on calls
}
\description{
get_existing_appointments()
Get a list of appointments already in calendar to avoid when creating clinics
}
\examples{
get_existing_appointments(dmy("1/1/17"), dmy("1/2/17"))
}
