% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainfuncs.r
\name{post_Sunday_clinics}
\alias{post_Sunday_clinics}
\title{post_Sunday_clinics()}
\usage{
post_Sunday_clinics(appts)
}
\arguments{
\item{appts}{A list of pre existing appointments created with get_existing_appointments()}
}
\description{
This function creates clinics on Monday afternoon only (not morning) after a Sunday on call
}
\examples{
# Using default parameters for get_existing_appointments()
post_sunday_clinics(get_existing_appointments())
}
