% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bankhols.r
\name{scrape_bh_dates}
\alias{scrape_bh_dates}
\title{Scrape bank holiday dates from UK government website}
\usage{
scrape_bh_dates()
}
\value{
A Date vector containing bank holidays for England
  for the current and next year
}
\description{
Scrape bank holiday dates from UK government website
}
\examples{
bh_list <- scrape_bh_dates()
}
