% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.r
\name{toRFC3339}
\alias{toRFC3339}
\title{toRFC3339()}
\usage{
toRFC3339(x)
}
\arguments{
\item{x}{A date-time object}
}
\value{
A string containing the date-time in an RFC 3339 compatible format
}
\description{
Convert lubridate time to RFC 3339 for use by google API
See https://tools.ietf.org/html/rfc3339
}
\details{
Assumes a UK location and corrects for British Summer Time if necessary
}
\examples{
toRFC3339(now())
toRFC3339(dmy("1/1/17"))

}
